requirement.txt
pytest
allure-pytest
selenium
Appium-Python-Client